Added fehCopiesCost.m script
PLEASE READ THE READ ME FILE TO UNDERSTAND THE SCRIPT. I am not sure if the script is correctly made. I try to generate a rank/stars to every summon in the summoning section. Afterwards a color. Then, for each summon with the same color I check the probability of that summon being the target unit of the rank/star of the summon. I end up using 3 random number generators. I also assume that MATLAB's rng is unbiased and that, for each pool, any unit is equiprobable to pull.